+++ printing/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Remove deprecated methods
 - Document.save() now returns a Future
+- Implement pan and zoom on PdfPreview widget
 
 ## 3.7.2